Experimental - separate server from the election logic
